beginner/submissions/team-members/konstantinos/requirements.txt
openai
tavily-python
python-dotenv
gradio
pydantic